requirements.txt
ccxt==4.3.74
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
python-telegram-bot==13.15
python-dotenv==1.0.1

